fix: ChannelPage - metadata유지 및 페이지 단일로 수정
