+++ src/features/channel/components/ChannelList/types.ts
@@ -1,7 +1,3 @@
-export interface ChannelListProps {
-  title: string;
-}
-
 export interface ChannelMapping {
   [key: string]: string[];
 }
